requirements.txt
Django==1.9.9
django-rest==0.0.1
djangorestframework==3.4.6
mysqlclient==1.3.7
